requirements.txt
pandas==1.4.4
plotly==5.9.0
numpy==1.21.5
inflection==0.5.1
streamlit==1.17.0
folium==0.14.0
Pillow==9.2.0
streamlit-folium==0.11.0
matplotlib==3.5.2


